Refactor event creation in CalendarEventsTest
Fixed the logic error in calendarGridContent_showsEventBlocks_whenEventsProvided. Previously it generated and event on the 14th of October and checked whether it was displayed; but since the calendar view dynamically changes to the current week, the 14th of October was not anymore in the span of the calendar. Fixing it by adding an event to a date of the current week.